# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "REST API minta backend",
        "version": "1.0.0",
        "description": "<img alt='DB_diagram' height='400px' src='https://nitslaszlo.github.io/static/shopking/db_diagram.jpg' />"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        },
        {
            "url": "https://shopking.cyclic.app"
        }
    ],
    "paths": {
...
        "/orderdetails": {
            "get": {
                "tags": ["orderdetails"],
                "summary": "Összes rendelésrészleteinek lekérdezése.",
                "description": "Az összes rendelés adatát tudjuk lekérdezni.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/orderdetails"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["orderdetails"],
                "summary": "Új rendelésrészlet mentése.",
                "description": "Az új rendelés mentése a 'orderdetails' táblában",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/orderdetails"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK (Az új, elmentett recepttel tér vissza.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/orderdetails"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/payments": {
            "get": {
                "tags": ["payments"],
                "summary": "Összes payment lekérdezése.",
                "description": "Az összes payment adatát tudjuk lekérdezni.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/payments"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["payments"],
                "summary": "Új payment mentése.",
                "description": "Az új payment mentése a 'payment' táblában",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/payments"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK (Az új, elmentett payment tér vissza.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/payments"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        {
            "name": []
        }
    ],
    "components": {
        "schemas": {
...
            "card": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "description": "A kártya neve.",
                        "example": "visa."
                    },
                    "lastFourNumbers": {
                        "type": "string",
                        "description": "A kártya utolso 4 számjegye.",
                        "example": "4252"
                    },
                    "expiryMonth": {
                        "type": "number",
                        "description": "A lejárati honap.",
                        "example": 12
                    },
                    "expiryYear": {
                        "type": "number",
                        "decription": "A tlejárati év",
                        "example": 2026
                    },
                    "cbbVerified": {
                        "type": "boolean",
                        "decription": "A kártya cbb kód érvényesége",
                        "example": true
                    }
                }
            },
...
            "orderdetails": {
                "type": "object",
                "required": ["archived_price", "discount", "quantity"],
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "PK - A rendelés azonosítója (ObjectID).",
                        "example": "63b53db2d18321524471f789"
                    },
                    "product_id": {
                        "type": "string",
                        "description": "FK - A termék id (ObjectID).",
                        "example": "6395d17cb8f84c79593f4292"
                    },
                    "order_id": {
                        "type": "string",
                        "description": "FK - A order id (ObjectID).",
                        "example": "6396d6fb9bd7d89ca308eb3f"
                    },
                    "archived_price": {
                        "type": "number",
                        "description": "A termék ára",
                        "example": 1900
                    },
                    "discount": {
                        "type": "number",
                        "description": "A termék leárazása",
                        "example": 200
                    },
                    "quantity": {
                        "type": "number",
                        "description": "A termék mennyisége",
                        "example": 1
                    }
                }
            },
            "payments": {
                "type": "object",
                "required": ["type", "status"],
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "PK - A rendelés azonosítója (ObjectID).",
                        "example": "63917f60978e764a7e2c6dd1"
                    },
                    "user_id": {
                        "type": "string",
                        "description": "FK - A termék id (ObjectID).",
                        "example": "63fc807e3f944579a10703b5"
                    },
                    "type": {
                        "type": "string",
                        "description": "A payment fajtája (credit, cash).",
                        "example": "credit"
                    },
                    "status": {
                        "type": "string",
                        "description": "A kártya érvényesége",
                        "example": "verified"
                    },
                    "card": {
                        "type": "object",
                        "$ref": "#/components/schemas/card",
                        "description": "A felhasználó kártya adatai.",
                        "example": true
                    }
                }
            },
...
            "ErrorSchema": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "error.message value"
                    },
                    "status": {
                        "type": "integer",
                        "example": 400
                    }
                }
            }
        }
    }
}